refactor(newsfeed): extract article categorization and filtering logic into separate utility functions
